requirements.txt
Crypto
Pillow